# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Pepinow",
        "description": "Pepinow Api Documentation",
        "version": "3.1.0"
    },
    "paths": {
        "/api/category": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Get all categories",
                "operationId": "index",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid status value"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Categories"
                ],
                "summary": "Store new category",
                "operationId": "store",
                "parameters": [
                    {
                        "name": "category",
                        "in": "query",
                        "description": "Category name to be stored",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "category created successfully"
                    },
                    "400": {
                        "description": "Invalid values"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/category/1": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Get category with Id",
                "operationId": "show",
                "responses": {
                    "201": {
                        "description": "category created successfully"
                    },
                    "400": {
                        "description": "Invalid values"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Categories"
                ],
                "summary": "Update category",
                "operationId": "update",
                "parameters": [
                    {
                        "name": "category",
                        "in": "query",
                        "description": "Category name to be stored",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Category updated successfully"
                    },
                    "400": {
                        "description": "Category not found"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Categories"
                ],
                "summary": "delete category",
                "operationId": "destroy",
                "responses": {
                    "201": {
                        "description": "category deleted successfully"
                    },
                    "400": {
                        "description": "Category not found"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/plant": {
            "get": {
                "tags": [
                    "Plants"
                ],
                "summary": "Get all Plants",
                "operationId": "get",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid status value"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Plants"
                ],
                "summary": "Store new Plant",
                "operationId": "storep",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Plant name to be stored",
                        "required": true
                    },
                    {
                        "name": "description",
                        "in": "query",
                        "description": "plant description to be stored",
                        "required": true
                    },
                    {
                        "name": "price",
                        "in": "query",
                        "description": "plant price to be stored",
                        "required": true
                    },
                    {
                        "name": "category_id",
                        "in": "query",
                        "description": "Plant category id to be stored",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid status value"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/plant/1": {
            "get": {
                "tags": [
                    "Plants"
                ],
                "summary": "Get plant",
                "operationId": "showp",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid status value"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Plants"
                ],
                "summary": "Update Plant",
                "operationId": "updatep",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Plant name to be updated",
                        "required": true
                    },
                    {
                        "name": "description",
                        "in": "query",
                        "description": "plant description to be updated",
                        "required": true
                    },
                    {
                        "name": "price",
                        "in": "query",
                        "description": "plant price to be updated",
                        "required": true
                    },
                    {
                        "name": "category_id",
                        "in": "query",
                        "description": "Plant category id to be updated",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid status value"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Plants"
                ],
                "summary": "delete Plant",
                "operationId": "destroyP",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid status value"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Login",
                "operationId": "authenticate",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "User email",
                        "required": true
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "User password",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid status value"
                    }
                }
            }
        },
        "/api/register": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Register",
                "operationId": "register",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "User name",
                        "required": true
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "User email",
                        "required": true
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "User password",
                        "required": true
                    },
                    {
                        "name": "password_confirmation",
                        "in": "query",
                        "description": "User password_confirmation",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid status value"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearer_token": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}